pyproject.toml
[project]
name = "aguitest"
version = "0.1.0"
description = "Learning AGUI (Agent User Interaction Protocol) SDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ag-ui-protocol>=0.1.9",
    "mumulib>=1.0.12",
    "pydantic-ai-slim[ag-ui,bedrock]>=1.1.0",
    "uvicorn>=0.38.0",
]
